Move aura into a feature renderer
Also pulled in a lib for it
