+++ build.gradle
@@ -10,6 +10,7 @@ repositories {
     maven { url "https://maven.shedaniel.me" }
     maven { url "https://maven.terraformersmc.com" }
     maven { url "https://ladysnake.jfrog.io/artifactory/mods" }
+    maven { url "https://maven.willbl.dev/" }
     maven { url "https://jitpack.io" }
 }
 
@@ -23,6 +24,7 @@ dependencies {
     modImplementation "dev.onyxstudios.cardinal-components-api:cardinal-components-entity:${project.cca_version}"
     modImplementation "com.terraformersmc:modmenu:${project.modmenu_version}"
     modImplementation "io.github.ladysnake:satin:${satin_version}"
+    modImplementation include("com.williambl.early_features:early_features:${project.early_features_version}")
 
     modApi("me.shedaniel.cloth:cloth-config-fabric:${project.cloth_config_version}") {
         exclude(group: "net.fabricmc.fabric-api")
