+++ gradle.properties
@@ -15,5 +15,6 @@ archives_base_name=arcanus
 fabric_version=0.47.8+1.18.2
 cca_version=4.1.4
 satin_version=1.7.2
+early_features_version=1.0.2
 cloth_config_version=6.0.42
 modmenu_version=3.0.0
