--- src/main/java/dev/cammiescorner/arcanus/client/AuraRenderer.java
@@ -1,25 +1,0 @@
-package dev.cammiescorner.arcanus.client;
-
-import net.minecraft.client.network.AbstractClientPlayerEntity;
-import net.minecraft.client.render.entity.EntityRenderer;
-import net.minecraft.client.render.entity.EntityRendererFactory;
-import net.minecraft.client.render.entity.LivingEntityRenderer;
-import net.minecraft.client.render.entity.model.PlayerEntityModel;
-import net.minecraft.entity.player.PlayerEntity;
-import net.minecraft.util.Identifier;
-
-public class AuraRenderer extends LivingEntityRenderer<AbstractClientPlayerEntity, PlayerEntityModel<AbstractClientPlayerEntity>> {
-    public AuraRenderer(EntityRendererFactory.Context ctx, PlayerEntityModel<AbstractClientPlayerEntity> model, float shadowRadius) {
-        super(ctx, model, shadowRadius);
-    }
-
-    /*
-     * TODO: create a biped model (or clone of player's model), inflate slightly, (maybe subdivide???), render with
-     *  aura shader with uniforms set.
-     */
-
-    @Override
-    public Identifier getTexture(AbstractClientPlayerEntity entity) {
-        return null;
-    }
-}
